backend/requirements.txt
fastapi[standard]
tortoise-orm
